maxsum

+++ lec 6/lec6.c
@@ -38,7 +38,7 @@ int main(){
 
 int main(){
     int n;                     // code for factorial
-    int fact=1 ,i;
+    long fact=1 ,i;
     printf("Enter number");
     scanf("%d",&n);
     for(i=1;i<=n;i++){
